Added user routes and blogpost routes. All routes need testing after DB seeding.

+++ controllers/blogpostRoutes.js
@@ -16,7 +16,7 @@ router.get('/', async (req,res) => {
     }
 });
 
-// TODO: get one blogpost | gotta be logged in to see details
+// Get one blogpost | gotta be logged in to see details
 router.get('/:id', authorizeCheck, async (req,res) => {
     try {
         const thisPost = await Blogpost.findByPk(req.params.id);
@@ -24,37 +24,48 @@ router.get('/:id', authorizeCheck, async (req,res) => {
             throw new Error('Blogpost could not be found');
         } else {
             const plainPost = thisPost.get( { plain: true }); // possibly might have to add an index of "[0]" before the .get
-            res.status(200).json(thisPost);
+            res.status(200).json(plainPost);
         };
     } catch (err) {
         res.status(500).json(err);
     };
 });
 
-// TODO: edit a blogpost | DEFINITELY gotta be logged in to edit a post, and it has to be YOUR post
+// Edit a blogpost | DEFINITELY gotta be logged in to edit a post, and it has to be YOUR post
 router.put('/:id', authorizeCheck ,async (req,res) => {
     try {
-        // TODO; make sure the blogpost's user_id property matches this user's id
-        Blogpost.update(req.body, {
-            where: {
-                id: req.params.id
-            }
-        });
-        res.status(200).json('post updated');
+        const thisPost = await (await Blogpost.findByPk(req.params.id)).get({ plain:true });
+
+        if (thisPost.user_id !== req.session.user_id) {
+            throw new Error('Edit blog post access not granted.')
+        } else {
+            Blogpost.update(req.body, {
+                where: {
+                    id: req.params.id
+                }
+            });
+            res.status(200).json('post updated');
+        };
     } catch (err) {
         res.status(500).json(err);
     };
 })
 
-// TODO: delete a blogpost | DEFINITELY the same requirement as editing to delete
+// Delete a blogpost | DEFINITELY the same requirement as editing to delete
 router.delete('/:id', authorizeCheck, async (req,res) => {
     try {
-        // TODO: check authorization
-        Blogpost.destroy({
-            where: {
-                id: req.params.is
-            }
-        })
+        const thisPost = await (await Blogpost.findByPk(req.params.id)).get({ plain:true });
+
+        if (thisPost.user_id !== req.session.user_id) {
+            throw new Error('Edit blog post access not granted.')
+        } else {
+            Blogpost.destroy({
+                where: {
+                    id: req.params.id
+                }
+            });
+            res.status(200).json(thisPost);
+        };
     } catch (err) {
         res.status(500).json(err);
     }
